requirement.txt
python-telegram-bot==13.7
psycopg2-binary==2.9.3
python-dotenv==0.19.2deactivate